added artist info screen

+++ app/src/main/java/com/example/musicwiki/network/model/Album.kt
@@ -22,5 +22,5 @@ data class Album(
     var attribute: Attribute? = null,
 
     @SerializedName("playcount")
-    var playCount: Long? = null
+    var playCount: Int? = null
 )
